+++ visualization-plot-gog/src/jvmTest/kotlin/visualization/plot/gog/plot/event3/TestUtil.kt
@@ -139,7 +139,7 @@ object TestUtil {
 
         val located = getGeomTargets(locator.findTargets(coord))
 
-        assertEquals(key.size.toLong(), located.size.toLong())
+        assertEquals(key.size, located.size)
 
         var i = 0
         val n = key.size
